backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
websockets>=11.0.3
pydantic>=2.8.0
python-multipart>=0.0.6
openai>=1.3.7
anthropic>=0.7.7
httpx>=0.25.2
python-json-logger>=2.0.7
aiofiles>=23.2.1
python-dotenv>=1.0.0
ollama>=0.3.0 
sqlalchemy>=2.0.0
aiosqlite>=0.19.0 